LeukoPatch V1
First commit of LeukoPatch. Intended to disinfect maya scenes and patch one of the viruses that spreads via maya scenes, (virus that originally was intended to target leukocyte but became the virus itself) and that infects userSetup.py file in maya/scripts folder. Only activates if both infection is detected and user consents to file patching.